Basic arg parsing and PDF support
Arguments are moving around so that OpenAI is the default but will fail
without OPENAI_API_KEY set. To run locally, use "--local". Eventually
users will be able to select the local HuggingFace model to run by
specifiying on the CL.

PDFReader Index has been added so the bot can contextualize to a PDF
provided by the filesystem.